pyproject.toml
[project]
name = "pmp-2"
version = "1.0.0"
description = "Панель мониторинга питанием для трех PZEM-016 220В 100А"
requires-python = ">=3.12"
dependencies = [
    "blinker==1.9.0",
    "click==8.1.8",
    "colorama==0.4.6",
    "Flask==3.1.0",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.5",
    "MarkupSafe==3.0.2",
    "paho-mqtt==2.1.0",
    "pymodbus==3.8.6",
    "Werkzeug==3.1.3"
]
[tool.setuptools]
py-modules = ["main"]  # Если есть main.py в корне